pyproject.toml
[tool.poetry]
name = "pywencai"
version = "0.4.2"
description = ""
authors = ["pluto <mayuanchi1029@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
PyExecJS = "^1.5.1"
requests = "*"
pandas = "^1.5.0"

[tool.poetry.group.dev.dependencies]
notebook = "^6.4.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
